Replace group_leader with our capturing StringIO
We tried this before but gave up on it.
Process.sleep was enough to guarantee we didn't see standard IO output
(".", "E", "?"..., etc.) from other tests when asking about diff. Now
that we've gotten rid of Process.sleep, we need it again.

This time we restore previous group_leader on process termination to
make sure we won't loose any output at the end of tests.